Funtion to create standard/wild decks added
The funtion to create standard/wild decks
was added to SelectClassNewDeckGUI.java,
CreateDeckGUI.java was modified to display
standard/wild cards when the user is creating
or editing a deck.
The class Deck.java now has a funtion to know
if the deck is standard or not.
2 buttons were added to TrackDeckGUI.java in
case of any player surrendered.